# docs/api/library-v1.openapi.yaml
openapi: 3.1.0
...
info:
  title: XiaDown Library API
  version: 1.0.0
  description: |
    Version 1 of the device-facing XiaDown Library API. The API is exposed only
    by a user-enabled Library transport. Pairing credentials and Bearer tokens
    are secrets; clients must not put them in URLs, analytics, logs, or crash
    reports.

    The pairing response grants `library.read` and `tasks.read` by default.
    `music.read`, `music.state`, `music.manage`, `rss.read`, `rss.state`,
    `rss.manage`, `rss.fetch`, `tasks.create`, and `tasks.control` can be granted
    later by the desktop owner. Pairing never grants either RSS management or
    public-feed fetch access by default.
    A grant is bound to one Catalog, so a valid token from another Catalog is
    rejected.

    A stable device ID and public-key hash are client assertions, not authority
    to inherit an existing grant. Pairing rejects collisions with an active
    grant. After the Desktop owner revokes a grant, a fresh short-lived pairing
    session may replace it in place: the grant identity and creation time remain
    for namespace continuity, but the token and public-key hash rotate and scopes
    reset to the default minimum. The old token remains invalid and prior managed
    scopes must be granted again by the Desktop owner.

    Canonical response property names are case-sensitive camelCase. Path and
    query parameter names are case-sensitive. Unless an input is explicitly
...
servers:
...
    variables:
...
security: []
...
paths:
...
  /api/v1/pair:
    post:
      tags: [Pairing]
      operationId: pairDevice
      summary: Exchange a short-lived pairing session for a device Bearer token
      description: |
        The pairing session is consumed only after its device grant is stored
        successfully. An active device-ID collision is rejected. A revoked
        identity may be replaced in place with a rotated credential and the
        default minimum scopes; no prior managed scope is inherited.
      requestBody:
        required: true
        x-max-body-bytes: 32768
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairRequest"
      responses:
        "201":
          description: Device paired. The returned token is shown only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairResponse"
        "400":
          description: Malformed JSON, unknown fields, an oversized body, invalid request fields, or a collision with an active device grant (`invalid_request`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: The pairing session is invalid, expired, exhausted, or already consumed (`pairing_invalid`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: The pairing exchange could not be completed (`pairing_failed`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/device-access:
    get:
      tags: [Pairing]
      operationId: getDeviceAccess
      summary: Read the current catalog-bound device grant and effective scopes
      description: This credential-introspection endpoint requires a valid Bearer token but no business-data scope.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The identity and effective scopes bound to the presented credential.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The credential belongs to a different Catalog (`catalog_access_denied`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/library:
    get:
      tags: [Library]
      operationId: getLibraryOverview
      summary: Read the default Catalog overview and synchronization cursor
      security:
        - bearerAuth: []
      x-xiadown-required-scope: library.read
      responses:
        "200":
          description: Current Catalog counts, health counts, and a high-water synchronization cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: xd1 opaque device token
      description: "Send `Authorization: Bearer <token>`. The token is Catalog-bound and must be stored as a secret."
...
  responses:
...
    Unauthorized:
      description: Missing, malformed, expired, revoked, or invalid Bearer token (`unauthorized`).
      headers:
        WWW-Authenticate:
          description: Bearer challenge for the XiaDown Library realm.
          schema:
            type: string
            const: 'Bearer realm="xiadown-library"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The token belongs to another Catalog (`catalog_access_denied`) or lacks the operation scope (`insufficient_scope`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Opaque server failure (`internal_error`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorCode:
      type: string
      description: Stable machine-readable API error code. Error messages and implementation details are intentionally absent.
      enum:
        - asset_stream_limit_reached
        - asset_unavailable
        - catalog_access_denied
        - insufficient_scope
        - idempotency_conflict
        - internal_error
        - invalid_cursor
        - invalid_music_mutation
        - invalid_pagination
        - invalid_request
        - music_content_changed
        - music_dependency_pending
        - music_entity_not_found
        - music_idempotency_conflict
        - music_revision_conflict
        - music_stream_limit_reached
        - not_found
        - pairing_failed
        - pairing_invalid
        - reset_required
        - rss_state_conflict
        - task_state_conflict
        - unauthorized
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorCode"
...
    DeviceScope:
      type: string
      enum:
        - library.read
        - music.read
        - music.state
        - music.manage
        - rss.read
        - rss.state
        - rss.manage
        - rss.fetch
        - tasks.read
        - tasks.create
        - tasks.control
    PairRequest:
      type: object
      additionalProperties: false
      required: [nonce, code, deviceID, name, publicKeyHash]
      properties:
        nonce:
          type: string
          minLength: 1
          description: Opaque nonce from the versioned pairing link.
        code:
          type: string
          pattern: "^[0-9]{6}$"
        deviceID:
          type: string
          minLength: 1
          maxLength: 255
          description: Stable, case-sensitive client installation identifier.
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: User-visible device name.
        publicKeyHash:
          type: string
          pattern: "^(?:sha256:)?[0-9A-Fa-f]{64}$"
          description: SHA-256 fingerprint of the device public key, optionally prefixed by `sha256:`.
    PairResponse:
      type: object
      additionalProperties: false
      required: [grantID, token, scopes]
      properties:
        grantID:
          type: string
          minLength: 1
          maxLength: 255
        token:
          type: string
          minLength: 1
          x-sensitive: true
          description: Opaque one-time-returned Bearer credential. Store it in protected device storage.
        scopes:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/DeviceScope"
    DeviceAccessResponse:
      type: object
      additionalProperties: false
      required: [grantId, catalogId, deviceId, scopes]
      properties:
        grantId:
          type: string
          minLength: 1
          maxLength: 255
        catalogId:
          type: string
          minLength: 1
          maxLength: 255
        deviceId:
          type: string
          minLength: 1
          maxLength: 255
        scopes:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/DeviceScope"
        capabilities:
          type: array
          uniqueItems: true
          description: |-
            Mounted shared protocol capabilities. `sync-events-v1` gates
            `GET /api/v1/sync/events`; clients must ignore unknown values.
          items:
            type: string
            minLength: 1
        stations:
          type: object
          additionalProperties: false
          description: Mounted device-facing Station APIs and this grant's authorization for each one.
          properties:
            library:
              $ref: "#/components/schemas/StationAccessSummary"
            music:
              $ref: "#/components/schemas/StationAccessSummary"
            rss:
              $ref: "#/components/schemas/StationAccessSummary"
    StationAccessSummary:
      type: object
      additionalProperties: false
      required: [supported, authorized, capabilities]
      properties:
        supported:
          type: boolean
          const: true
        authorized:
          type: boolean
        capabilities:
          type: array
          uniqueItems: true
          description: |-
            Mounted capability identifiers. The RSS Station currently advertises
            `rss-sync-v1`, `opaque-resource-slots-v1`,
            `rss-subscription-mutations-v1`, `rss-shared-public-fetch-v1`,
            `rss-observations-v1`, and `rss-fetch-lease-v1`. Clients must still
            require the corresponding top-level scope and ignore unknown values.
          items:
            type: string
            minLength: 1
...
    CatalogStatus:
      type: string
      enum: [active, archived]
...
    Catalog:
      type: object
      additionalProperties: false
      required: [id, name, status, isDefault, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/CatalogStatus"
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryCounts:
      type: object
      additionalProperties: false
      required: [all, video, audio, books, images, others]
      properties:
        all:
          type: integer
          minimum: 0
        video:
          type: integer
          minimum: 0
        audio:
          type: integer
          minimum: 0
        books:
          type: integer
          minimum: 0
        images:
          type: integer
          minimum: 0
        others:
          type: integer
          minimum: 0
    StatusCounts:
      type: object
      additionalProperties: false
      required: [active, needsReview, missing, trashed]
      properties:
        active:
          type: integer
          minimum: 0
        needsReview:
          type: integer
          minimum: 0
        missing:
          type: integer
          minimum: 0
        trashed:
          type: integer
          minimum: 0
    HealthCounts:
      type: object
      additionalProperties: false
      required:
        - assetLinks
        - itemsWithoutAssets
        - unavailableAssetFiles
        - legacyFilesWithErrors
        - offlineStorageRoots
        - readOnlyStorageRoots
        - storageRootsWithErrors
      properties:
        assetLinks:
          type: integer
          minimum: 0
        itemsWithoutAssets:
          type: integer
          minimum: 0
        unavailableAssetFiles:
          type: integer
          minimum: 0
        legacyFilesWithErrors:
          type: integer
          minimum: 0
        offlineStorageRoots:
          type: integer
          minimum: 0
        readOnlyStorageRoots:
          type: integer
          minimum: 0
        storageRootsWithErrors:
          type: integer
          minimum: 0
    SyncState:
      type: object
      additionalProperties: false
      required: [epoch, cursor]
      properties:
        epoch:
          type: string
        cursor:
          type: integer
          format: int64
          minimum: 0
    LibraryOverview:
      type: object
      additionalProperties: false
      required: [catalog, categories, statuses, health, sync, capabilities]
      properties:
        catalog:
          $ref: "#/components/schemas/Catalog"
        categories:
          $ref: "#/components/schemas/CategoryCounts"
        statuses:
          $ref: "#/components/schemas/StatusCounts"
        health:
          $ref: "#/components/schemas/HealthCounts"
        sync:
          $ref: "#/components/schemas/SyncState"
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [snapshot-keyset-v1]
...